requirements.txt
pandas
numpy
django-ninja
lightgbm
gunicorn